yolov8-detection/requirements.txt
viam-sdk
ultralytics
torch
torchvision
pillow
numpy
opencv-python
